# health_service/[OLD]health_interface/swagger_server/models/correct_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CorrectResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, command: str=None, description: str=None):  # noqa: E501
        """CorrectResponse - a model defined in Swagger

        :param command: The command of this CorrectResponse.  # noqa: E501
        :type command: str
        :param description: The description of this CorrectResponse.  # noqa: E501
        :type description: str
        """
        self.swagger_types = {
            'command': str,
            'description': str
        }

        self.attribute_map = {
            'command': 'command',
            'description': 'description'
        }
        self._command = command
        self._description = description

    @classmethod
    def from_dict(cls, dikt) -> 'CorrectResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CorrectResponse of this CorrectResponse.  # noqa: E501
        :rtype: CorrectResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def command(self) -> str:
        """Gets the command of this CorrectResponse.

        type of results: ok or error  # noqa: E501

        :return: The command of this CorrectResponse.
        :rtype: str
        """
        return self._command

    @command.setter
    def command(self, command: str):
        """Sets the command of this CorrectResponse.

        type of results: ok or error  # noqa: E501

        :param command: The command of this CorrectResponse.
        :type command: str
        """

        self._command = command

    @property
    def description(self) -> str:
        """Gets the description of this CorrectResponse.

        Brief description of the operation result  # noqa: E501

        :return: The description of this CorrectResponse.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this CorrectResponse.

        Brief description of the operation result  # noqa: E501

        :param description: The description of this CorrectResponse.
        :type description: str
        """

        self._description = description